pyproject.toml
[project]
name = "stt-and-translate-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.46.0",
    "python-dotenv>=1.1.1",
    "sarvamai>=0.1.21",
    "tiktoken>=0.12.0",
]
